[#1570] moved S3 logic to s3RestApi
