+++ data/src/test/java/org/akvo/flow/data/repository/FormDataRepositoryTest.java
@@ -31,6 +31,7 @@ import org.akvo.flow.data.entity.form.FormIdMapper;
 import org.akvo.flow.data.entity.form.XmlFormParser;
 import org.akvo.flow.data.net.RestApi;
 import org.akvo.flow.data.net.s3.AmazonAuthHelper;
+import org.akvo.flow.data.net.s3.S3RestApi;
 import org.akvo.flow.data.util.ApiUrls;
 import org.akvo.flow.domain.util.DeviceHelper;
 import org.junit.After;
@@ -106,18 +107,21 @@ public class FormDataRepositoryTest {
     private MockWebServer mockWebServer;
     private FormDataRepository formDataRepository;
     private RestApi restApi;
+    private S3RestApi s3RestApi;
 
     @Before
     public void setUp() throws IOException {
         mockWebServer = new MockWebServer();
         mockWebServer.start(8080);
 
+        ApiUrls apiUrls = new ApiUrls(null, null);
         restApi = spy(new RestApi(mockDeviceHelper, new TestRestServiceFactory(),
-                "1.2.3", new ApiUrls(null, null), mockAmazonAuth, mockDateFormat, null));
+                "1.2.3", apiUrls));
+        s3RestApi = spy(new S3RestApi(new TestRestServiceFactory(), apiUrls, mockAmazonAuth, mockDateFormat, null));
         DataSourceFactory dataSourceFactory = new DataSourceFactory(null, null,
                 mockDatabaseDataSource, null, mockFileDataSource, null);
         formDataRepository = new FormDataRepository(mockFormHeaderParser, mockXmlParser,
-                restApi, dataSourceFactory, mockFormIdMapper);
+                restApi, dataSourceFactory, mockFormIdMapper, s3RestApi);
 
         when(mockFormHeaderParser.parseOne(anyString())).thenReturn(mockApiFormHeader);
         when(mockFormHeaderParser.parseMultiple(anyString())).thenReturn(
@@ -165,7 +169,7 @@ public class FormDataRepositoryTest {
 
         observer.assertNoErrors();
         observer.assertValueCount(1);
-        verify(restApi, times(0)).downloadArchive(anyString());
+        verify(s3RestApi, times(0)).downloadArchive(anyString());
         verify(mockFileDataSource, times(0))
                 .extractRemoteArchive(any(ResponseBody.class), anyString());
     }
@@ -189,7 +193,7 @@ public class FormDataRepositoryTest {
 
         observer.assertNoErrors();
         observer.assertValueCount(1);
-        verify(restApi, times(1)).downloadArchive(anyString());
+        verify(s3RestApi, times(1)).downloadArchive(anyString());
         verify(mockFileDataSource, times(1))
                 .extractRemoteArchive(any(ResponseBody.class), anyString());
     }
